Remove HTTP/JSON leak from Game.Core domain Extensions (#1142)
Game.Core.Extensions carried helpers that pulled HTTP/codegen concerns
into the domain layer. Resolved per the issue:

- Removed ToBase64<T> and Deserialize<T>(HttpResponseMessage): both were
  dead (only their own tests referenced them); removing the latter drops
  the System.Net.Http coupling from the domain project.
- Relocated Decapitalize and SnakeCase to Game.Api/CodeGen's
  CodeGenExtensions, their sole consumers. (The issue guessed these were
  unused; they are codegen-only string helpers, so they move rather than
  delete.)
- Capitalize/SpaceWords stay in Game.Core (used by domain models and
  reference-data seeding).

Tests moved alongside the methods.
